Only report Changed:True on actual changes
Add --stats to rsync command and parse the output to report Change:True
only if rsync's output contains 'deleting' and don't contain
'transferred: 0' which means it didn't do any transfers because the
files were the same.
It requires verbosity=1 to work in case there are no new files but only
deletions.